refactor(application): extract shared idea access validation into extension method in idea
