+++ src/Backend/Services/Polls/Polls.Application/Ideas/Commands/UpdateIdea/UpdateIdeaCommandHandler.cs
@@ -3,7 +3,6 @@ using MediatR;
 using Polls.Application.Common.Interfaces;
 using Polls.Application.Ideas.DTOs;
 using Polls.Application.Ideas.Guards;
-using Polls.Application.Polls.Guards;
 using Polls.Domain.Common;
 using Polls.Domain.Ideas;
 
@@ -24,18 +23,9 @@ public sealed class UpdateIdeaCommandHandler(
 
         if (!command.BypassRestrictions)
         {
-            var ideaGuard = IdeaGuard.For(idea)
-                .IsOwner(command.UserId)
-                .IsNotApproved()
-                .Validate();
-            if (!ideaGuard.IsSuccess)
-                return ideaGuard.Error;
-
-            var guardResult = PollGuard.For(idea.Poll)
-                .IsNotFinished()
-                .Validate();
-            if (!guardResult.IsSuccess)
-                return guardResult.Error;
+            var validationResult = idea.ValidateIdeaAccess(command.UserId);
+            if (!validationResult.IsSuccess)
+                return validationResult.Error;
         }
 
         idea.Title = command.Title;
